+++ src/initial/initial_field_coefficients_planewave.m
@@ -53,7 +53,7 @@ alpha = simulation.input.initialField.azimuthalAngle;
 [pilm,taulm] = spherical_functions_trigon(cb,sb,lmax);  % Nk x 1
 
 % cylindrical coordinates for relative particle positions
-relativeParticlePositions = bsxfun(@plus,simulation.input.particles.positionArray,-simulation.input.initialField.focalPoint);
+relativeParticlePositions = simulation.input.particles.positionArray-simulation.input.initialField.focalPoint;
 kvec = k*[sb*cos(alpha);sb*sin(alpha);cb];
 eikr = exp(1i*relativeParticlePositions*kvec);