+++ src/postprocessing/initial_field_plane_wave_pattern.m
@@ -70,7 +70,7 @@ switch simulation.input.initialField.type
             emnikrg = exp(-1i*(kxGrid*RG(1) + kyGrid*RG(2) + kzGrid*RG(3)));  % Nalpha x Nbeta
             prefacCosBetGaussfac = E0*k^2*w^2/(4*pi)* cos(betaArray(:).') .* exp(-w^2/4*k^2*sin(betaArray(:).').^2)  .*  ( sign(cos(betaArray(:).')) == sign(cos(initF.polarAngle)) ) ;  % 1 x Nbeta, make sure only the correct direction contributes
 
-            eikrgPrefacCosBetGaussfac = bsxfun(@times,emnikrg,prefacCosBetGaussfac); % Nalpha x Nbeta
+            eikrgPrefacCosBetGaussfac = emnikrg.*prefacCosBetGaussfac; % Nalpha x Nbeta
 
             switch lower(simulation.input.initialField.polarization)
                 case 'te'
@@ -79,8 +79,8 @@ switch simulation.input.initialField.type
                     alphaG = simulation.input.initialField.azimuthalAngle-pi/2;
             end
             
-            pwp{1}.expansionCoefficients = bsxfun(@times, cos(alphaArray(:)-alphaG), eikrgPrefacCosBetGaussfac);
-            pwp{2}.expansionCoefficients = sign(cos(initF.polarAngle)) * bsxfun(@times, sin(alphaArray(:)-alphaG), eikrgPrefacCosBetGaussfac);
+            pwp{1}.expansionCoefficients = cos(alphaArray(:)-alphaG).*eikrgPrefacCosBetGaussfac;
+            pwp{2}.expansionCoefficients = sign(cos(initF.polarAngle)) * (sin(alphaArray(:)-alphaG).*eikrgPrefacCosBetGaussfac);
 
         else
             error('this is not implemented')